app/requirements.txt
debugpy==1.8.5
exif==1.6.0
gps==3.19
keyboard==0.13.5
numpy==1.24.0
Pillow==10.4.0
pymysql==1.1.1
pynmea2==1.19.0
pyserial==3.5
simple_pid==2.0.1
av==10.0.0
picamera2==0.3.19
